# PizzaTaskDeliveryMS/src/main/java/com/example/pizzataskms/model/OrderRequest.java
package com.example.pizzataskms.model;

import java.util.Objects;

import com.fasterxml.jackson.annotation.JsonProperty;
import com.example.pizzataskms.model.Order;
import com.fasterxml.jackson.annotation.JsonCreator;
import java.util.ArrayList;
import java.util.List;
import javax.validation.constraints.*;

@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-01-27T12:36:56.679+02:00")

public class OrderRequest   {
  @JsonProperty("orders")
  private List<Order> orders = new ArrayList<Order>();

  public OrderRequest orders(List<Order> orders) {
    this.orders = orders;
    return this;
  }

  public OrderRequest addOrdersItem(Order ordersItem) {
    this.orders.add(ordersItem);
    return this;
  }

  public List<Order> getOrders() {
    return orders;
  }

  public void setOrders(List<Order> orders) {
    this.orders = orders;
  }


  @Override
  public boolean equals(java.lang.Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    OrderRequest orderRequest = (OrderRequest) o;
    return Objects.equals(this.orders, orderRequest.orders);
  }

  @Override
  public int hashCode() {
    return Objects.hash(orders);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class OrderRequest {\n");
    
    sb.append("    orders: ").append(toIndentedString(orders)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(java.lang.Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }
}